Begin to decouple query fragments from SQLite
Currently, query fragments are modeled as strings with `?`s denoting
bindings, and an array of bindings. Instead we should model things as a
single array of segments that can be concatenated with whatever form of
binding identifier we need. For example, Postgres can use `$1`, `$2`,
etc.

This remodel also has the positive effect of safer debug descriptions of
our queries.